Themes: redirect logged-in-only routes to login page in logged-out sessions
For theme uploads, there can be only one route definition for both logged-in and logged-out
sessions. When logged out, the `redirectLoggedOut` handler will kick in and will redirect to
login page.

For other `/themes` routes, we redirect the logged-in-only ones (i.e., with `:site_id` suffix)
to login page when hit in logged-out session.