Redirect sections that don't have the enableLoggedOut flag to login page
Routes in sections that don't have the `enableLoggedOut` flag set to `true` should always
redirect to login page when requested in a logged-out session.

This patch removes the check-and-redirect code (which is broken anyway) from the `loggedOutMiddleware`
function and moves it to the section loader handler. Which, until now, only loaded the section's
webpack chunk. Now it also guards for unwanted (logged-out) visitors.